Add support for dynamic macros
This change will detect if an error is thrown for a missing module, if this
error is detected we will try to traverse the file and download all remaining
dependencies recursively. We then sync this back to the manager so new babel
workers will have these files as well.

This comes down to having a mini bundler and file fetcher for every babel
worker to execute babel macros.

Fixes #981 #1266