Show a loader on the button while forking
This is a follow-up of #1667.
The effect is simple because for now I didn't want to modify the markup of the button.

As far as my limited CSS skills lead me, @CompuIves suggestion regarding nzbin.github.io/three-dots would require to have a div wrapper around the button and a div to display the 3 dots. Do we want to do this?