✨ Add Frozen/Editing Indicator, Apply Coding Conventions to Header (WIP)
Refactored a bunch of the Header components for better organization and to conform to coding conventions.

Added a new HazardButton primitive and used it to create a ToggleFrozenButton for the Sandbox Header. This button has the same effect as the Frozen toggle switch inside of the Project Info workspace and is used to give the user a visual indicator that they are viewing a frozen or editable sandbox.

TODO:
- Maybe move the HazardButton to common components and improve it's styling API. Could also use better types.
- Change the Fork button to show 'Use Template" when the current Sandbox is a Template. This would create a new Sandbox using the Template, but would sever the link to source GitHub repository, which the current Fork method preserves.
- Maybe rearrange the Header buttons?
- Continue refactoring remaining Header components to apply coding conventions.