🔨 Refactor app/components/SubscribeForm: Apply Conventions
Updated react-stripe-elements from ^v1.7 to ^v3.0.0

injectStripe doesn't appear to want to expose the CheckoutForm props and I'm not sure if there is a way I can work around it. Someone with better TypeScript knowledge than me will have to figure this one out.

Refactored all components to apply coding conventions. SubrscribeForm is basically unchanged, but CheckoutForm was re-written using hooks, so it needs to be tested to ensure that it still works.