🧹 Cleanup PathedTemplates
Applied coding conventions and added types.

It's worth noting that I was unable to add types for react-router-dom without having to upgrade multiple dependencies, which quickly grew outside of the scope of this PR. So instead proper types will be added at a later date once the build process and project dependencies are overhauled.