Fix babel warning when downloading dynamic files in worker
Whenever we try to download dynamic files in a worker we need to write the
downloaded files to the filesystem. The reason for this is that we have a
separate fs in the worker that's synced eagerly, because when we need to execute
files in the worker we can only retrieve the files synchronously. We used to
save the files to the WorkerFS, the WorkerFS is responsible for downloading the
files from the main thread. But this FS would get confused by also writing to it
from the worker, and then get out of sync.

I now wrapped the WorkerFS with an OverlayFS so that writes are written on top of
the FS, but not to the FS. I had to bump our browserfs version for this.