Add ownership to GitHub sandboxes
This allows users to change properties of a sandbox if they're the owner.
They can change the title, description, things like env variables. Everything
except the code, in which we'll autofork.

The goal of this is to answer to the multiple feature requests related to using
a git sandbox. While the best solution would be to invent a permission model,
we're not there yet and I'm not sure when we'll get it. So this seems like a good
intermediate solution instead.