Make the client cache a jwt token to prevent spamming the server (#5020)
The client requests for every socket reconnect a new token
This happens every 100ms at the start, and sometimes
the server hasn't responded before retrying, which causes a big
spam for the server. This change makes sure that we only request
a new token when we're sure that the existing request has finished.