Don't delete files with similar name optimistically
We used to return everything starting with the name of a directory in
`getModulesAndDirectoriesInDirectory`. This isn't right, because if you
have this:

- /app
- /app/index.js
- /app-components

`getModulesAndDirectoriesInDirectory(..., ..., '/app')` would also
include `/app-components`. Luckily we only do this client side, so when
someone deletes the folder `/app`, `/app-components` doesn't get
deleted. But it's confusing that the folder gets hidden. This fixes that