feat: make vite work in cra template
Our CRA bundler works already very similar to Vite, out of the box. Similar configuration, static files in the same place, transpilation logic is the same.
The only thing that is missing, is knowing that conventially `src/main.tsx` is the entrypoint.

With this change, we make the basic examples of Vite work, using our existing CRA bundler. When things get more advanced (Vite plugins, etc) it won't work, but then people can opt to export it to a Devbox. Until then, this should handle 90% of all cases.